requirements.txt
aws-glue-libs @ git+https://github.com/awslabs/aws-glue-libs # Glue 4.0
boto3 ==1.24.70 # From https://docs.aws.amazon.com/glue/latest/dg/aws-glue-programming-python-libraries.html#glue-modules-provided
pyspark ==3.3.0 # https://docs.aws.amazon.com/glue/latest/dg/release-notes.html
